Send 403 if contact does not belong to user

+++ delete.php
@@ -2,6 +2,8 @@
 
 require "database.php";
 
+session_start();
+
 if (!isset($_SESSION["user"])) {
   header ("Location: login.php");
   return;
@@ -17,6 +19,14 @@ if ($statement->rowCount() == 0){
   echo ("HTTP 404 NOT FOUND");
   return;
 } 
+
+$contact = $statement->fetch(PDO::FETCH_ASSOC);
+
+if ($contact["user_id"] != $_SESSION["user"]["id"]){
+  http_response_code (403);
+  echo("HTTP 403 UNAUTHORIZED");
+  return;
+}
 
 $conn->prepare("DELETE FROM contacts WHERE id = :id")->execute([":id" => $id]);
